+++ tests/test_gdpr_api_query.py
@@ -73,7 +73,7 @@ def test_gdpr_query_requires_correct_scope(
         assert response.status_code == 403
 
 
-def test_if_profile_not_found_return_404(
+def test_if_profile_not_found_return_204(
     api_client, user, uuid_value, requests_mock, settings
 ):
     auth_header = get_api_token_for_user_with_scopes(
@@ -87,4 +87,4 @@ def test_if_profile_not_found_return_404(
         )
     )
 
-    assert response.status_code == 404
+    assert response.status_code == 204