+++ tests/snapshots/snap_test_gdpr_api_query.py
@@ -21,3 +21,19 @@ snapshots["test_get_profile_information_from_gdpr_api 1"] = {
     ],
     "key": "PROFILE",
 }
+
+snapshots["test_model_lookup_can_be_configured_to_a_field 1"] = {
+    "children": [
+        {"key": "MEMO", "value": "Memo"},
+        {
+            "children": [
+                {"key": "FIRST_NAME", "value": "First name"},
+                {"key": "LAST_NAME", "value": "Last name"},
+            ],
+            "key": "USER",
+        },
+        {"key": "USER", "value": "First name Last name"},
+        {"children": [], "key": "EXTRA_DATA"},
+    ],
+    "key": "PROFILE",
+}