+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- `GDPR_API_MODEL_LOOKUP` setting for configuring how the GDPR model instance is found.
+
 ### Changed
 
 - Respond with a `204 No Content` status code when no data is found for the requested profile.